+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Cache only 2xx responses, closes #104
 - Set caching headers on statics, closes #103
 - Drop sensor TTLs, closes #102
+- Display sensor reading count
 
 ## `0.70.3`